requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
python-chess>=1.10.0
requests>=2.31.0
SpeechRecognition>=3.10.0
pydub>=0.25.1 